+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerExecutionAdapter.java
@@ -1,0 +1,14 @@
+package com.interface21.webmvc.servlet.mvc.tobe;
+
+import com.interface21.webmvc.servlet.ModelAndView;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+public class HandlerExecutionAdapter implements HandlerAdapter {
+
+    @Override
+    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler)
+            throws Exception {
+        return ((HandlerExecution) handler).handle(request, response);
+    }
+}
